deprecated/backend/requirements.txt
fastapi[standard]
sentence_transformers
numpy 
faiss-cpu